Split memory module into separate files
Structure this similarly to the CPU module, with separate files
for proc values and for each cgroup version.
